+++ src/main/server/api/v1/apple/mappings.ts
