+++ src/main/server/api/v1/apple/scripts.ts
