+++ src/main/server/api/v1/interfaces/attachmentInterface.ts
