+++ src/main/server/api/v1/interfaces/chatInterface.ts
