+++ src/main/server/api/v1/interfaces/generalInterface.ts
