+++ src/main/server/api/v1/interfaces/serverInterface.ts
