+++ src/main/server/services/http/api/v1/routers/serverRouter.ts
@@ -1,8 +1,8 @@
 import { Next } from "koa";
 import { RouterContext } from "koa-router";
 import { FileSystem } from "@server/fileSystem";
-import { GeneralInterface } from "../interfaces/generalInterface";
-import { ServerInterface } from "../interfaces/serverInterface";
+import { ServerInterface } from "@server/api/v1/interfaces/serverInterface";
+import { GeneralInterface } from "@server/api/v1/interfaces/generalInterface";
 import { Success } from "../responses/success";
 
 export class ServerRouter {
