+++ src/main/server/services/http/api/v1/routers/settingsRouter.ts
@@ -2,7 +2,7 @@ import { RouterContext } from "koa-router";
 import { Next } from "koa";
 
 import { isNotEmpty } from "@server/helpers/utils";
-import { BackupsInterface } from "../interfaces/backupsInterface";
+import { BackupsInterface } from "@server/api/v1/interfaces/backupsInterface";
 import { ValidateInput, ValidateJSON } from "../validators";
 import { SettingsValidator } from "../validators/settingsValidator";
 import { Success } from "../responses/success";
