+++ src/main/server/services/http/constants.ts
@@ -1,3 +1,2 @@
 export const MessagesBasePath = `${process.env.HOME}/Library/Messages`;
-
 export const invisibleMediaChar = String.fromCharCode(65532);
