+++ src/main/server/services/http/index.ts
@@ -14,15 +14,13 @@ import * as http from "http";
 import { Server } from "@server/index";
 import { FileSystem } from "@server/fileSystem";
 
-// Helpers
-import { createServerErrorResponse } from "@server/helpers/responses";
-
 // Entities
 import { EventCache } from "@server/eventCache";
 
 import { HttpRoutes as HttpRoutesV1 } from "./api/v1/httpRoutes";
 import { SocketRoutes as SocketRoutesV1 } from "./api/v1/socketRoutes";
 import { ErrorMiddleware } from "./api/v1/middleware/errorMiddleware";
+import { createServerErrorResponse } from "./api/v1/responses";
 
 /**
  * This service class handles all routing for incoming socket
