--- src/main/server/services/http/responses.ts
@@ -1,3 +1,0 @@
-export const BadRequest = ({ message = "Validation Error", errors = [] }: { message: string; errors: string[] }) => {
-    return { status: 400, message, errors };
-};
